+++ src/main/java/com/mineastr/MineAstr.java
@@ -29,7 +29,7 @@ import org.slf4j.Logger;
 @Mod(MineAstr.MODID)
 public final class MineAstr {
     public static final String MODID = "mineastr";
-    public static final String MOD_VERSION = "0.6.29";
+    public static final String MOD_VERSION = "0.6.30";
     public static final Logger LOGGER = LogUtils.getLogger();
 
     private static final MineAstrBridge BRIDGE = new MineAstrBridge();